Make get_cfps be precise in the amount of memory it copies
Instead of trying to get 100 frames, possibly going beyond the base of
the stack, this finds the base of the stack and only copies from the
current frame to the base.

This is necessary to make it work with the coredump, where 100 frames
back from the cfp is not actually present.