Report C functions in the Ruby stack as "unknown C function"
Previously we were leaving C functions out of the stack that rbspy
reports entirely. A most correct way to handle this is to instead report
them as "unknown C function" -- this still doesn't give much insight
into what the unknown C function **is** exactly.

This was causing an issue with profiling Ruby HTTP servers (for instance
Sinatra), which sometimes spend a lot of time in the `select` syscall
with no other Ruby function running.