Handle execve's on Linux
When an execve is executed, the process's memory is replaced with the memory
of a new program. When the new program is also a Ruby interpreter instance,
rbspy should continue to gather stack traces. This is particularly important
because some Ruby invocation methods, such as rvm, may do an execve to launch
the script being invoked by the user.

The strategy here is to check the result returned when getting the stack trace
from the target process to detect the error that the pointer into the ruby
thread state, specifically, has become invalid. If it has, we then reinitialize
StackTraceGetter on the basis that if we've lost our thread the whole process
is invalid.

Currently this doesn't work on MacOS because the error-handling for copying
addresses can't distinguish between an invalid read and process exit.
Fixing that would require either changing the interface of
read-process-memory to surface the mach return codes or just working around
it in rbspy.

It might also not work as reliably on 32-bit platforms, where there may be
a greater chance of the thread state pointer being remapped into the
memory of the new process.