Fix Windows sleep time
On windows, the sleep call could take up to 15ms irregardless of the amount of
time asked for. On my system without the sleep call, I can get about 14K stack
traces a second - but with the sleep call we were limited to 64 causing an
error message to display when sampling.

Fix by using timeBeginPeriod call to make the OS to wake up every 1ms. This lets
us sample up with rates up to 500hz without errors.