Refactor subprocess test to coordinate with the spawned rubies
Instead of relying on timing, which is problematic in CI, this approach lets the ruby processes run until they're told (by the rbspy test code) that their stack frame has been received and that they can exit. This has the side benefit of making the test faster, as we're not waiting an arbitrary amount of time for ruby to exit on its own.