Retry all operations in get_process_ruby_state to avoid race conditions
The race conditions mostly seem to affect Windows but can occur on other platforms too. The old assumption seemed to be that subsequent symbol lookups would work after we'd found the ruby version symbol, but that isn't always true. We now retry the entire batch of lookups until we have everything that we need or we've exhausted the number of allowed retries.

This also improves our error messaging when we fail to get the state of the ruby process. #268