Handle permission errors while trying to lock
If a permission error is encountered while trying to lock the target process,
returning ProcessEnded is a misrepresentation of the error: the process
has not ended, we failed to lock it due to a permission error.

This matters because ProcessEnded is the error condition that leads to a
successful exit, so permission errors may result in a premature exit.

This commit also fixes a logical error in test_spawn_record_children_subprocesses
which could result in the test itself causing an EPERM issue to arise on Linux,
as it would cause the root process that it spawns to be defunct, leading to
EPERM while trying to use ptrace to lock it.